+++ package/sdk/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.4.0",
-  genVersion: "2.500.5",
-  userAgent: "speakeasy-sdk/typescript 0.4.0 2.500.5 1.0.0 ryan-total-test-act",
+  sdkVersion: "0.4.1",
+  genVersion: "2.502.0",
+  userAgent: "speakeasy-sdk/typescript 0.4.1 2.502.0 1.0.0 ryan-total-test-act",
 } as const;