+++ package/sdk/RELEASES.md
@@ -288,3 +288,13 @@ Based on:
 - [typescript v0.4.0] package/sdk
 ### Releases
 - [NPM v0.4.0] https://www.npmjs.com/package/ryan-total-test-act/v/0.4.0 - package/sdk
+
+## 2025-02-04 00:08:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.483.0 (2.502.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.1] package/sdk
+### Releases
+- [NPM v0.4.1] https://www.npmjs.com/package/ryan-total-test-act/v/0.4.1 - package/sdk